系统功能调整

+++ z-module-system/src/main/java/com/z/module/system/domain/Department.java
@@ -3,7 +3,9 @@ package com.z.module.system.domain;
 import com.z.framework.common.domain.AbstractAuditingEntity;
 import lombok.Data;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
 import java.io.Serializable;
@@ -15,7 +17,7 @@ public class Department extends AbstractAuditingEntity implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private Integer parentId;
+    private Long parentId;
 
     @NotEmpty(message = "部门名称不能为空")
     @Size(max = 50, message = "部门名称长度不能超过50")
